Move link logic to cspell-lib
